Remove DependencyInjection dll.
My initial intention was to allow RdClient package to be used w/ and w/o a DI container. However it appeared to be quite difficult to support. Now RdClient can be used with DI container only, so it now makes sense to put DI-related code into the same assembly and remove separate DependencyInjection dll.